Modify the Version api to accept security options.
This allows accessing an ssl secure endpoint to check its version using
the provided options.
Currently auth is supported only for etcd.
Modify the Version implementation in etcd to use the crypto package to
parse the provided certificate files. It uses http.Get("https://<ip>/version")
to get the version.