PWX-9227 PX panics with 'invalid memory address or nil pointer dereference' when consul cluster goes down
The meta values is returned as nil when consul loses quorum, this causes
a SegFault when we try to dereference the meta pointer to get the lastIndex.
This, was happening because we were overwriting the error in the client.go code
where we retryConnection(), thus ignoring the actual error that we got while accessing consul.

Signed-off-by: Tapas Sharma <tapas@portworx.com>